requirements.txt
Django==4.1.3
psycopg2==2.9.4
ccxt==1.95.24
django-apscheduler==0.6.2
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.13.0
django-sslserver==0.22
djangorestframework==3.14.0
python-dotenv==0.6.2
requests==2.28.1
marshmallow==3.19.0
yarl==1.7.2
rsa==4.9
django-rest-swagger==2.1.2
djangorestframework-simplejwt





